adding all for an analysis on local

+++ jsolver.py
@@ -137,11 +137,11 @@ def dF_dv2(Fu, Fv):
 # helper functions to pack/unpack and reshape the solutions
 @jit
 def pack_electrostatic(V, C):
-    return jnp.array(jnp.concatenate((jnp.ravel(V), jnp.ravel(C))), dtype = jnp.float64)
+    return jnp.array(jnp.concatenate((jnp.ravel(V), jnp.ravel(C))))
 
 @jit
 def pack_magnetostatic(V, Fu, Fv, C):
-    return jnp.array(jnp.concatenate((jnp.ravel(V), jnp.ravel(Fu), jnp.ravel(Fv), jnp.ravel(C))), dtype = jnp.float64)
+    return jnp.array(jnp.concatenate((jnp.ravel(V), jnp.ravel(Fu), jnp.ravel(Fv), jnp.ravel(C))))
 
 @jit
 def unpack_electrostatic(V_C):
@@ -210,7 +210,7 @@ def newton(f, x_0, tol=tol, max_iter=max_iter):
 if inputfile is None:
     # use bulk solution as initial guess for the electrostatic problem and perform Newton's method
     V0 = jnp.full((NU, NV), -1)
-    C0 = jnp.full((NU, NV), jnp.sqrt(-J), dtype = jnp.float64)
+    C0 = jnp.full((NU, NV), jnp.sqrt(-J))
     x0 = pack_electrostatic(V0, C0)
     start = time.time()
     electrostatic_solution = newton(f_electrostatic, x0)
